requirements.txt
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mathparse==0.1.2
mysql-connector==2.2.9
mysqlclient==2.0.3
nltk==3.5
numpy==1.20.1
packaging==20.9
Pillow==8.1.1
Pillow-PIL==0.1.dev0
Pint==0.16.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.7.5
PyYAML==3.13
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.2.19
tqdm==4.58.0
Werkzeug==1.0.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
requests==2.21.0
urllib3==1.24.1
